+++ lib/option.dart
@@ -16,6 +16,7 @@ library option;
 
 import 'dart:async';
 
+import 'result.dart';
 import 'src/util.dart';
 
 part 'src/option/option.dart';